refactor: move avatar file validation to zod-schemas
